Refactor create order tests
