+++ e2e/tests/testCACaseCreator.spec.ts
@@ -5,7 +5,7 @@ test.use({ storageState: Config.sessionStoragePath + "solicitor.json" });
 
 test.describe("CA Case creation examples", (): void => {
   test("create case and submit and pay", async ({
-    manageCasesEventUtils
+    manageCasesEventUtils,
   }): Promise<void> => {
     await manageCasesEventUtils.submitTSSolicitorCase("C100");
   });
